refactor is_user_admin to common and remove it from base view
