+++ superset/dashboards/filters.py
@@ -24,10 +24,11 @@ from sqlalchemy import and_, or_
 from sqlalchemy.orm.query import Query
 
 from superset import db, is_feature_enabled, security_manager
+from superset.common.request_contexed_based import is_user_admin, get_user_roles
 from superset.models.core import FavStar
 from superset.models.dashboard import Dashboard
 from superset.models.slice import Slice
-from superset.views.base import BaseFilter, get_user_roles, is_user_admin
+from superset.views.base import BaseFilter
 from superset.views.base_api import BaseFavoriteFilter
